Testes correção de bugs

+++ spring-matutino/src/main/java/br/com/trier/springmatutino/services/impl/PaisServiceImlp.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import br.com.trier.springmatutino.domain.Pais;
+import br.com.trier.springmatutino.domain.User;
 import br.com.trier.springmatutino.repositories.PaisRepository;
 import br.com.trier.springmatutino.services.PaisService;
 import br.com.trier.springmatutino.services.exceptions.ObjetoNaoEncontrado;
@@ -30,7 +31,7 @@ public class PaisServiceImlp implements PaisService {
 	@Override
 	public Pais findById(Integer id) {
 		Optional<Pais> obj = repository.findById(id);
-		return obj.orElseThrow(() -> new ObjetoNaoEncontrado("País %s não encontrado".formatted(id)));
+		return obj.orElseThrow(() -> new ObjetoNaoEncontrado("Usuário %s não encontrado".formatted(id)));
 	}
 
 	@Override
@@ -41,11 +42,14 @@ public class PaisServiceImlp implements PaisService {
 	@Override
 	public void delete(Integer id) {
 		Pais pais = findById(id);
+		if (pais != null) {
+			throw new ObjetoNaoEncontrado("País %s não encontrado".formatted(id));
+		}
 		repository.delete(pais);
 	}
 
 	@Override
-	public List<Pais> findByNameIgnoreCase(String name) {
+	public List<Pais> findByName(String name) {
 		List<Pais> lista = repository.findByName(name);
 		if (lista.size() == 0) {
 			throw new ObjetoNaoEncontrado("Nenhum nome de país inicia com %s".formatted(name));
